# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CLARK Gateway",
    "version": "3.7.6",
    "description": "The API gateway for all of the CLARK microservices",
    "license": { "name": "ISC", "url": "https://www.isc.org/licenses/" },
    "contact": {
      "name": "SecurEd Inc.",
      "url": "https://secured.team/",
      "email": "info@secured.team"
    },
    "x-logo": {
      "url": "https://clark.center/assets/images/logo.png",
      "altText": "CLARK Logo"
    },
    "x-tagGroups": [
      { "name": "Library Service", "tags": ["Library Service"] },
      { "name": "Notification Service", "tags": ["Notification Service"] },
      { "name": "Feature Service", "tags": ["Feature Service"] },
      { "name": "Rating Service", "tags": ["Rating Service"] },
      { "name": "Outcome Service", "tags": ["Outcome Service"] },
      {
        "name": "Learning Object Service",
        "tags": ["Learning Object Service"]
      },
      { "name": "User Service", "tags": ["User Service"] },
      { "name": "Utility Service", "tags": ["Utility Service"] },
      {
        "name": "Standard Guideline Service",
        "tags": ["Standard Guideline Service"]
      }
    ]
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Development" },
    { "url": "https://api-gateway.clark.center", "description": "Production" }
  ],
  "paths": {
...
    "/users/{userId}/learning-objects/{cuid}/changelog": {
      "post": {
        "description": "Creates a changelog",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The username of the author"
          },
          {
            "in": "path",
            "name": "cuid",
            "schema": { "type": "string" },
            "required": true,
            "description": "The cuid of the learning object"
          }
        ],
        "requestBody": {
          "description": "The cuid, version, and author's username",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "changelogText": {
                    "type": "string",
                    "description": "This is the text of the changelog",
                    "required": true,
                    "example": "This is an example"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is not a admin or editor"
          },
          "404": {
            "description": "NOT FOUND - Learning object or user not found"
          }
        }
      }
    },
    "/users/{userId}/learning-objects/{cuid}/changelogs": {
      "get": {
        "description": "Gets a changelog for a object",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The username of the author"
          },
          {
            "in": "path",
            "name": "cuid",
            "schema": { "type": "string" },
            "required": true,
            "description": "The cuid of the learning object"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Changelog"
                }
              }
            }
          }
        }
      }
    },
...
    "/collections/{name}/meta": {
      "get": {
        "description": "Gets a collection's meta data",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "schema": { "type": "string" },
            "required": true,
            "description": "The name of the collection"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Collection" }
              }
            }
          },
          "404": { "description": "NOT FOUND - Collection not found" }
        }
      }
    },
...
    "/learning-objects/{learningObjectId}/collections": {
      "patch": {
        "description": "Adds object to collection",
        "tags": ["Learning Object Service"],
        "requestBody": {
          "description": "The collection",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "collection": {
                    "type": "string",
                    "example": "nccp",
                    "description": "The collection name",
                    "required": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "UNAUTHENTICATED - User is not logged in" }
        }
      }
    },
    "/users/{username}/learning-objects/{cuid}/collection": {
      "patch": {
        "description": "Updates a in review object's collection",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The username of the author"
          },
          {
            "in": "path",
            "name": "cuid",
            "schema": { "type": "string" },
            "required": true,
            "description": "The cuid of the object"
          }
        ],
        "requestBody": {
          "description": "The collection",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "collection": {
                    "type": "string",
                    "example": "nccp",
                    "description": "The collection name",
                    "required": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "BAD REQUEST - Collection not defined in request body"
          },
          "401": { "description": "UNAUTHENTICATED - User is not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is not an admin or editor"
          },
          "404": {
            "description": "NOT FOUND - Learning object or collection not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Changelog": {
        "properties": {
          "cuid": {
            "type": "string",
            "description": "The cuid relating to the changelog",
            "example": "00000000-1111-2222-3333-444444444444",
            "required": true
          },
          "logs": {
            "type": "array",
            "description": "The changes made to the object",
            "required": true,
            "items": { "$ref": "#/components/schemas/ChangelogLog" }
          }
        }
      },
      "ChangelogLog": {
        "properties": {
          "date": {
            "type": "number",
            "description": "The date the changelog was created in ms",
            "example": 0,
            "required": true
          },
          "text": {
            "type": "string",
            "description": "The changelog text",
            "example": "This is a change",
            "required": true
          },
          "author": {
            "type": "object",
            "$ref": "#/components/schemas/ChangelogAuthor"
          }
        }
      },
      "ChangelogAuthor": {
        "properties": {
          "userId": {
            "type": "string",
            "description": "The id of the changelog author",
            "example": 0,
            "required": true
          },
          "name": {
            "type": "string",
            "description": "The name of the changelog author",
            "example": "John Doe",
            "required": true
          },
          "role": {
            "type": "string",
            "description": "The role of the changelog author",
            "example": "admin",
            "required": true
          },
          "profileImage": {
            "type": "string",
            "description": "The profile image of the changelog author",
            "required": true
          }
        }
      },
...
      "Collection": {
        "properties": {
          "_id": {
            "type": "string",
            "example": 0,
            "required": true,
            "description": "The collection id"
          },
          "name": {
            "type": "string",
            "example": "nccp",
            "required": true,
            "description": "The collection name"
          },
          "fullName": {
            "type": "string",
            "example": "National Cybersecurity Curriculum Program",
            "required": false,
            "description": "The collection's full name"
          },
          "abvName": {
            "type": "string",
            "example": "nccp",
            "required": true,
            "description": "The collection's abbreviated name"
          },
          "hasLogo": {
            "type": "boolean",
            "example": true,
            "required": true,
            "description": "True if the collection has a logo"
          },
          "description": {
            "type": "string",
            "example": "This is a collection",
            "required": true,
            "description": "Description of the collection"
          }
        }
      },
...
    }
  },
...
}